+++ HISTORY.md
@@ -44,6 +44,7 @@ Complete list of issues solved in the 1.0.0 milestone:
   security reasons, enable by setting [debug][db_log_queries] = 1 in testswarm.ini.
 * (#142) Implement database update script.
 * (#158) Implement database install script.
+* Implement browserset query script.
 
 ### User agents